requirements.txt
APScheduler==3.10.1
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.2.0
coloredlogs==15.0.1
ddddocr==1.4.7
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
itchat-uos==1.5.0.dev0
mpmath==1.3.0
numpy==1.25.1
onnxruntime==1.15.1
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==9.5.0
protobuf==4.23.4
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sympy==1.12
tzlocal==5.0.1
urllib3==2.0.4
